reorganizes codebase and adds lombok annotations to admin
