+++ src/main/java/sai_adapa/projs/inv_management/auth/AdminVerifierInterceptor.java
@@ -1,15 +1,15 @@
-package sai_adapa.projs.inv_management.auth.user_authorization;
+package sai_adapa.projs.inv_management.auth;
 
 import org.springframework.beans.factory.annotation.Configurable;
 import org.springframework.web.servlet.HandlerInterceptor;
+import sai_adapa.projs.inv_management.services.AdminService;
 import sai_adapa.projs.inv_management.tools.SpringContext;
-import sai_adapa.projs.inv_management.users.admin.AdminService;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 @Configurable
-public class AdminVerifier implements HandlerInterceptor {
+public class AdminVerifierInterceptor implements HandlerInterceptor {
 
 
     private AdminService getAdminService() {
