+++ src/main/java/sai_adapa/projs/inv_management/auth/UsersVerifierInterceptor.java
@@ -1,21 +1,20 @@
-package sai_adapa.projs.inv_management.auth.user_authorization;
+package sai_adapa.projs.inv_management.auth;
 
 import org.springframework.web.servlet.HandlerInterceptor;
+import sai_adapa.projs.inv_management.services.UsersService;
 import sai_adapa.projs.inv_management.tools.SpringContext;
-import sai_adapa.projs.inv_management.users.user.UsersService;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-public class UsersVerifier implements HandlerInterceptor {
+public class UsersVerifierInterceptor implements HandlerInterceptor {
 
     private UsersService getUsersService() {
         return SpringContext.getBean(UsersService.class);
     }
 
     @Override //Todo: add error resp in case of failure
-    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler)
-    {
+    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) {
         return getUsersService().verifySession(request.getHeader("session_token"));
     }
 }
