+++ src/main/java/sai_adapa/projs/inv_management/config/WebConfig.java
@@ -1,12 +1,11 @@
 package sai_adapa.projs.inv_management.config;
 
-import org.apache.catalina.User;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import sai_adapa.projs.inv_management.auth.user_authorization.AdminVerifier;
-import sai_adapa.projs.inv_management.auth.user_authorization.UsersVerifier;
-import sai_adapa.projs.inv_management.auth.user_authorization.VendorVerifier;
+import sai_adapa.projs.inv_management.auth.AdminVerifierInterceptor;
+import sai_adapa.projs.inv_management.auth.UsersVerifierInterceptor;
+import sai_adapa.projs.inv_management.auth.VendorVerifierInterceptor;
 
 
 @Configuration
@@ -16,9 +15,9 @@ public class WebConfig implements WebMvcConfigurer {
     public void addInterceptors(InterceptorRegistry registry)
     {
 
-        registry.addInterceptor(new AdminVerifier()).addPathPatterns("/admin","/admin/**").excludePathPatterns("/admin/new","/admin/login");
-        registry.addInterceptor(new UsersVerifier()).addPathPatterns("/app_user/logout");
-        registry.addInterceptor(new VendorVerifier()).addPathPatterns(("/vendor/logout"));
+        registry.addInterceptor(new AdminVerifierInterceptor()).addPathPatterns("/admin","/admin/**").excludePathPatterns("/admin/new","/admin/login");
+        registry.addInterceptor(new UsersVerifierInterceptor()).addPathPatterns("/app_user/logout");
+        registry.addInterceptor(new VendorVerifierInterceptor()).addPathPatterns(("/vendor/logout"));
     }
 
 }
