+++ src/main/java/sai_adapa/projs/inv_management/controllers/AdminController.java
@@ -1,12 +1,14 @@
-package sai_adapa.projs.inv_management.users.admin;
+package sai_adapa.projs.inv_management.controllers;
 
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
-import sai_adapa.projs.inv_management.users.user.PreUsers;
-import sai_adapa.projs.inv_management.users.user.UsersService;
-import sai_adapa.projs.inv_management.users.vendor.PreVendor;
-import sai_adapa.projs.inv_management.users.vendor.VendorService;
+import sai_adapa.projs.inv_management.services.AdminService;
+import sai_adapa.projs.inv_management.users.io.PreAdmin;
+import sai_adapa.projs.inv_management.users.io.PreUsers;
+import sai_adapa.projs.inv_management.services.UsersService;
+import sai_adapa.projs.inv_management.users.io.PreVendor;
+import sai_adapa.projs.inv_management.services.VendorService;
 
 @RestController
 public class AdminController {
