+++ src/main/java/sai_adapa/projs/inv_management/controllers/UsersController.java
@@ -1,6 +1,8 @@
-package sai_adapa.projs.inv_management.users.user;
+package sai_adapa.projs.inv_management.controllers;
 
 import org.springframework.web.bind.annotation.*;
+import sai_adapa.projs.inv_management.users.io.PreUsers;
+import sai_adapa.projs.inv_management.services.UsersService;
 
 @RestController
 public class UsersController {
