+++ src/main/java/sai_adapa/projs/inv_management/repositories/VendorRepository.java
@@ -1,7 +1,8 @@
-package sai_adapa.projs.inv_management.users.vendor;
+package sai_adapa.projs.inv_management.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+import sai_adapa.projs.inv_management.users.Vendor;
 
 @Repository
 public interface VendorRepository extends JpaRepository<Vendor,String > {
