+++ src/main/java/sai_adapa/projs/inv_management/services/UsersService.java
@@ -1,9 +1,11 @@
-package sai_adapa.projs.inv_management.users.user;
+package sai_adapa.projs.inv_management.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
-import sai_adapa.projs.inv_management.auth.AuthTools;
+import sai_adapa.projs.inv_management.tools.AuthTools;
+import sai_adapa.projs.inv_management.repositories.UsersRepository;
+import sai_adapa.projs.inv_management.users.io.PreUsers;
+import sai_adapa.projs.inv_management.users.Users;
 
 @Service
 public class UsersService {
@@ -29,12 +31,12 @@ public class UsersService {
         return usersRepository.findByEmail(e_mail);
     }
 
-    @Cacheable(value = "usersFromSession", key = "#token")
+
     public Users getUsersBySession(String token) {
         return usersRepository.findBySessionToken(token);
     }
 
-    @Cacheable(value = "usersSessionStatus", key = "#token")
+
     public Boolean verifySession(String token) {
         return usersRepository.existsUsersBySessionToken(token);
     }
