+++ src/main/java/sai_adapa/projs/inv_management/services/VendorService.java
@@ -1,9 +1,11 @@
-package sai_adapa.projs.inv_management.users.vendor;
+package sai_adapa.projs.inv_management.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
-import sai_adapa.projs.inv_management.auth.AuthTools;
+import sai_adapa.projs.inv_management.tools.AuthTools;
+import sai_adapa.projs.inv_management.repositories.VendorRepository;
+import sai_adapa.projs.inv_management.users.io.PreVendor;
+import sai_adapa.projs.inv_management.users.Vendor;
 
 @Service
 public class VendorService {
@@ -38,12 +40,12 @@ public class VendorService {
         vendorRepository.save(vendor);
     }
 
-    @Cacheable(value = "vendorFromSession", key = "#token")
+
     public Vendor getUserBySession(String token) {
         return vendorRepository.findBySessionToken(token);
     }
 
-    @Cacheable(value = "vendorSessionStatus", key = "#token")
+
     public Boolean verifySession(String token) {
         return vendorRepository.existsVendorBySessionToken(token);
     }
