+++ src/main/java/sai_adapa/projs/inv_management/tools/AuthTools.java
@@ -1,4 +1,4 @@
-package sai_adapa.projs.inv_management.auth;
+package sai_adapa.projs.inv_management.tools;
 
 
 import org.springframework.security.crypto.password.Pbkdf2PasswordEncoder;
@@ -10,7 +10,7 @@ public class AuthTools {
 
     private static final SecureRandom secureRandom = new SecureRandom(); //threadsafe
     private static final Base64.Encoder base64Encoder = Base64.getUrlEncoder(); //threadsafe
-    private static Pbkdf2PasswordEncoder pbkdf2PasswordEncoder;
+    private static final Pbkdf2PasswordEncoder pbkdf2PasswordEncoder;
 
     static {
         pbkdf2PasswordEncoder = new Pbkdf2PasswordEncoder();
