+++ src/main/java/sai_adapa/projs/inv_management/tools/RequestLogger.java
@@ -1,9 +1,10 @@
 package sai_adapa.projs.inv_management.tools;
 
 import org.springframework.context.annotation.Configuration;
+import org.springframework.stereotype.Component;
 import org.zalando.logbook.Logbook;
 
-@Configuration
+@Component
 public class RequestLogger {
     Logbook logbook = Logbook.create();
 }
