+++ src/main/resources/application.properties
@@ -1,7 +1,3 @@
 logging.level.org.zalando.logbook= TRACE
 logbook.format.style=json
 logging.file.name= X:/spring_logs/inv.logs
-spring.cache.type=redis
-spring.redis.host=localhost
-spring.redis.port=6379
-spring.cache.redis.time-to-live=90
